move to query/dummy_catalog
